stemtivistas_backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stemtivistas-backend"
version = "1.0.0"
description = "Flask REST API backend for Stemtivistas web application"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Stemtivistas Team"}
]
dependencies = [
    "Flask>=3.0.0",
    "Flask-CORS>=4.0.0",
    "Flask-Bcrypt",
    "psycopg2-binary", # Use "PyMySQL" if you have a MySQL database
    "PyJWT",
    "python-dotenv",
    "sqlalchemy",
    "boto3",
    "gunicorn",
    "uvicorn",
    "fastapi",
    "passlib",
    "pydantic",
    "typing",
    "email_validator",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[tool.setuptools]
packages = []
